[core] core-rest-pipeline: redeclare RestError as interfaces (#34591)
### Packages impacted by this PR

- `@azure/core-rest-pipeline`

### Describe the problem that is addressed by this PR

We are currently doing some
[hackery](https://github.com/timovv/azure-sdk-for-js/blob/1b5b2014fecaedab9eb675af513a7fa2351c01ce/sdk/core/core-rest-pipeline/src/restError.ts#L76-L81)
to make constructing RestError from core-rest-pipeline yield the same
type as a RestError from `@typespec/ts-http-runtime`.

The main reason for doing this was that the two RestError types are
subtly incompatible (deep down in PipelineRequest, we moved from
AbortSignalLike to native AbortSignal), preventing a direct re-export.
This was the best I thought we could do to prevent having two different
kinds of RestError around while keeping the API surface as a class. But
this approach created some funky behavior when doing an instanceof check
-- if you call `instanceof` using the core-rest-pipeline RestError on
the right hand side, it will return false always (using the
ts-http-runtime RestError here works fine in all cases).

Here I am proposing that we don't do that hacky thing and instead expose
RestError and its constructor as an interface instead of as a class.
This allows us to reuse the ts-http-runtime definition directly without
the weirdness. I _think_ that these type definitions are compatible with
the original class, but would be great if others could take a look and
confirm -- are there any edge cases I'm missing where this approach
might cause issues?

### What are the possible designs available to address the problem? If
there are more than one possible design, why was the one in this PR
chosen?

- We could also override `Symbol.hasInstance` in the core-rest-pipeline
RestError instead, but that feels like we're just leaning into the hack
approach even more and it could expose us to more strangeness.
- Otherwise, we could accept the break in the type surface and just
re-export RestError from ts-http-runtime directly. To me this feels
risky as making changes like this in Core tends to cause unexpected
compile errors, even if there's no difference at runtime (we've been
here before).
- As a third option, we just accept that there are two different kinds
of `RestError`. `instanceof` will not be guaranteed to work, and so it
will break people. But maybe we just draw a line in the sand and say
"don't use instanceof". The `isRestError` type guard will continue to
work fine.