First cut at a test recording agent (#34755)
### Packages impacted by this PR

- Test recorder

### Describe the problem that is addressed by this PR

Playing around with a custom prompt for the test recorder. I have tried
it out on Key Vault, where I messed with the sanitizer configuration to
see if it could figure out what the problem was. It did, and re-added
the global sanitizer that I removed. I was quite impressed.

I'm hoping that we can give this a go, see if it is helpful, and further
refine it as folks try it out. Hopefully it will save some time and get
better as time goes on.

---------

Co-authored-by: Harsha Nalluru <harsha.nalluru1997@gmail.com>