[Monitor OpenTelemetry] Support `OTEL_METRIC_EXPORT_INTERVAL` for Metric Export Configuration (#34929)
### Packages impacted by this PR
@azure/monitor-opentelemetry

### Issues associated with this PR
#34837

### Describe the problem that is addressed by this PR
MetricHandler now prioritizes the `OTEL_METRIC_EXPORT_INTERVAL`
environment variable for setting the export interval, falling back to
options.collectionInterval or a default if not set or invalid. Added
comprehensive unit tests to verify correct precedence and handling of
various environment variable values.

### Are there test cases added in this PR? _(If not, why?)_
Yes

### Checklists
- [x] Added impacted package name to the issue description
- [ ] Does this PR needs any fixes in the SDK Generator?** _(If so,
create an Issue in the
[Autorest/typescript](https://github.com/Azure/autorest.typescript)
repository and link it here)_
- [x] Added a changelog (if necessary)