[dev-tool] bring back --force option for samples publish command (#35069)
If `--force` is passed to `samples publish` command, we log a warning instead of throwing an error.

This option was removed in the refactoring PR https://github.com/Azure/azure-sdk-for-js/pull/18556 but no reason is provided.
In fact, we still have `--force` and `-f` usage in our NPM scripts today and they are not doing anything.

This PR brings `--force` option back so that we can use it in intended scenarios (for example, code-gen automation)

Related issue:
https://github.com/Azure/azure-sdk-tools/issues/10938#issuecomment-3011673979