Call automation ga5 changes (#34590)
At present, CallAutomation uses the alpha API in the main branch, while
GA (General Availability) releases are managed from a separate feature
branch. With this update, GA changes will be merged into the main
branch, and all future GA releases will be published from there. Ongoing
alpha development will be moved to the **callautomation/feature/alpha**
branch

**Note:** Recordings under the callautomation are specific to call
automation. The assets.json has request and response that is pushed to a
remote repository and the events specific to call automation are sitting
in the call automation folder recordings. This should not be deleted as
it is needed for live test

---------

Co-authored-by: v-pivamshi <155710746+v-pivamshi@users.noreply.github.com>
Co-authored-by: v-saasomani <v-saasomani@microsoft.com>