[@azure/cosmos] Fix ENOTFOUND retry (#35113)
### Packages impacted by this PR
@azure/cosmos

### Issues associated with this PR
#32044

### Describe the problem that is addressed by this PR
This PR addresses the problem of Client retrying 120*120 times on an
inaccessible endpoint.

### What are the possible designs available to address the problem? If
there are more than one possible design, why was the one in this PR
chosen?
Now defaultRetryPolicy is applied when ENOTFOUND is the error code.


### Are there test cases added in this PR? _(If not, why?)_


### Provide a list of related PRs _(if any)_


### Command used to generate this PR:**_(Applicable only to SDK release
request PRs)_

### Checklists
- [ ] Added impacted package name to the issue description
- [ ] Does this PR needs any fixes in the SDK Generator?** _(If so,
create an Issue in the
[Autorest/typescript](https://github.com/Azure/autorest.typescript)
repository and link it here)_
- [ ] Added a changelog (if necessary)