[Cosmos] Enhance partition key extraction logic and update tests for system key (#34765)
### Packages impacted by this PR
@azure/cosmos

### Issues associated with this PR

### Describe the problem that is addressed by this PR
This PR fixes a bug where the partition key for a migrated container was
incorrectly returned as an empty array ([]). It now returns the correct
partition key value.

### What are the possible designs available to address the problem? If
there are more than one possible design, why was the one in this PR
chosen?


### Are there test cases added in this PR? _(If not, why?)_
Yes

### Provide a list of related PRs _(if any)_


### Command used to generate this PR:**_(Applicable only to SDK release
request PRs)_

### Checklists
- [ ] Added impacted package name to the issue description
- [ ] Does this PR needs any fixes in the SDK Generator?** _(If so,
create an Issue in the
[Autorest/typescript](https://github.com/Azure/autorest.typescript)
repository and link it here)_
- [ ] Added a changelog (if necessary)

---------

Co-authored-by: Manik Khandelwal <mkhandelwal@microsoft.com>