[core-util] attempt to fix flaky test (#35079)
Another theory is that `setTimeout` timer is not precise and 50ms may be within
the margin of error. This PR makes function1 delay and function2 delay ten times
of their old values so that the 500ms difference between the half of function1
delay and function1 delay is much larger to ensure that `aborter.abort()` is
invoked by the timeout callback before function 1 resolves.

Revert previous change to disable file parallelism as it didn't fix the issue.