[identity] Avoid waiting for app to close (#35406)
When using InteractiveBrowserCredential we set `wait` to true, which
forces the promise to wait until the app is _closed_ before fulfilling the promise.
This isn't quite right, as we want to hand control back to MSAL as soon as
the app opens and MSAL will then listen to the auth code and resolve the promise
with an access token.

The reason we did not notice this was because according to the 
[open documentation](https://github.com/sindresorhus/open?tab=readme-ov-file#wait)
you have to explicitly specify an app for it to be able to wait which we do
not do.

As a result, on Mac, IBC does not resolve correctly until the browser
instance is exited. 

This PR removes the `wait` override, ensuring that control is handed
back to MSAL as soon as the browser opens.

Manual validation:

|Platform|Browser|Tested|
|-----|-----|-----|
|Windows|Chromium|x|
|Windows|Safari||
|Windows|Firefox|x|
|OSX|Chromium|x|
|OSX|Safari|x|
|OSX|Firefox|x|


Fixes #33641