[playwright][lint] Ensure __dirname is used correctly (#35553)
When dual-emitting CJS and ESM, certain concepts are no longer available
to be used. One such example that keeps coming up is `__dirname`. Another may
be `__filename`. Neither is available in an ESM context.

While TSHY does support correctly polyfilling these, it is based on
convention and can be easy to miss.

This linter rule ensures that only files that end in -cjs.cts can
include these as identifiers. In the future we can enhance this to add other
options.

Fixes #35532
Resolves #35549