chore(loadtesting): removed useCloudHostedBrowsers (#35579)
### Packages impacted by this PR

@azure/create-playwright
@azure/playwright


### Issues associated with this PR

The useCloudHostedBrowsers option previously allowed users to choose
whether they wanted to use cloud-hosted browsers or not. This is no
longer necessary because if customers do not want to use cloud-hosted
browsers, they can simply run the tests using their local configuration.
Therefore, we are removing this option.

### Describe the problem that is addressed by this PR


### What are the possible designs available to address the problem? If
there are more than one possible design, why was the one in this PR
chosen?


### Are there test cases added in this PR? _(If not, why?)_


### Provide a list of related PRs _(if any)_


### Command used to generate this PR:**_(Applicable only to SDK release
request PRs)_

### Checklists
- [ ] Added impacted package name to the issue description
- [ ] Does this PR needs any fixes in the SDK Generator?** _(If so,
create an Issue in the
[Autorest/typescript](https://github.com/Azure/autorest.typescript)
repository and link it here)_
- [ ] Added a changelog (if necessary)