[mcp server] remove mcp server package (#35622)
### Packages impacted by this PR

- `@azure-tools/mcp-server`

### Describe the problem that is addressed by this PR

We are implementing a cross-language MCP server hosted inside the tools
repo
[here](https://github.com/Azure/azure-sdk-tools/tree/main/tools/azsdk-cli).
The idea is that partners will only have to install a single MCP server
for all the languages (and in fact this server is already set up in our
mcp.json). This combined MCP server will shell out to dev-tool for the
JS-specific tasks that we have.

Since all languages are rallying around this combined MCP server, we no
longer need the skeleton that we scaffolded out a while back, so this PR
deletes it. If we ever change direction here, we can always resurrect
it.