[Monitor OpenTelemetry Exporter] Customer SDK Stats Fixes (#35828)
### Packages impacted by this PR
@azure/monitor-opentelemetry-exporter

### Describe the problem that is addressed by this PR
This pull request introduces a new drop reason for disk persistence
disablement in the customer SDK Stats, refines the handling of drop and
retry reasons, and updates related enums and logic for improved clarity
and consistency.

**Customer SDK Stats:**
* Added a new drop reason, `CLIENT_STORAGE_DISABLED`, to track when disk
persistence is disabled for telemetry.
[[1]](diffhunk://#diff-994a7fbc4177f96448b6098ea5c95fc5872615b8537e18822a17817cd3d70870R16)
[[2]](diffhunk://#diff-2f4066f1034ba4adc1b1010a2db050dd609f6b2f1d5935ac9e0574d293c148f1R378-R379)
[[3]](diffhunk://#diff-88b62eef1c76c2db9a03f6c670881c41bab6de71bbb9c3fceeb1c33f0c9a455dL248-R257)

**Enums and Reason Handling:**
* Updated the `DropReason` enum to include `CLIENT_STORAGE_DISABLED` and
changed the default unknown value to "Unknown reason" for clarity.
* Introduced a new `RetryReason` enum to standardize retry reason
strings, replacing hardcoded values in the code.
[[1]](diffhunk://#diff-2f4066f1034ba4adc1b1010a2db050dd609f6b2f1d5935ac9e0574d293c148f1L12-R19)
[[2]](diffhunk://#diff-2f4066f1034ba4adc1b1010a2db050dd609f6b2f1d5935ac9e0574d293c148f1L526-R538)
[[3]](diffhunk://#diff-88b62eef1c76c2db9a03f6c670881c41bab6de71bbb9c3fceeb1c33f0c9a455dL248-R257)
* Refactored logic in `CustomerSDKStatsMetrics` to use the new
`RetryReason` enum for retry reason descriptions.

These changes improve the accuracy of telemetry drop and retry reporting
and make the codebase easier to maintain.

### What are the possible designs available to address the problem? If
there are more than one possible design, why was the one in this PR
chosen?

### Checklists
- [x] Added impacted package name to the issue description
- [ ] Does this PR needs any fixes in the SDK Generator?** _(If so,
create an Issue in the
[Autorest/typescript](https://github.com/Azure/autorest.typescript)
repository and link it here)_
- [x] Added a changelog (if necessary)