[Identity] Update VSCodeCred Behavior + Broker Docs (#35837)
### Packages impacted by this PR
@azure/identity
@azure/identity-vscode
@azure/identity-broker

### Describe the problem that is addressed by this PR
There's a scenario in which interactive authentication might show up for
`VisualStudioCodeCredential` when the plugin is set but the broker is
not available. This is because in MSAL implementation, a stub class
instance is imported so that static imports won't throw.

The PR ensures:
* `disableAutomaticAuthentication` is set internally for
`VisualStudioCodeCredential` so no interactive authentication would be
required
* Our SDK only proceeds when the plugin is configured and the broker is
available. We'll throw errors when either of those is not met with
distinct error message for easier debugging.
* Add additional troubleshooting and prerequisite for clearer doc

### What are the possible designs available to address the problem? If
there are more than one possible design, why was the one in this PR
chosen?


### Are there test cases added in this PR? _(If not, why?)_


### Provide a list of related PRs _(if any)_


### Command used to generate this PR:**_(Applicable only to SDK release
request PRs)_

### Checklists
- [ ] Added impacted package name to the issue description
- [ ] Does this PR needs any fixes in the SDK Generator?** _(If so,
create an Issue in the
[Autorest/typescript](https://github.com/Azure/autorest.typescript)
repository and link it here)_
- [ ] Added a changelog (if necessary)

---------

Co-authored-by: Scott Addie <10702007+scottaddie@users.noreply.github.com>
Co-authored-by: Charles Lowell <10964656+chlowell@users.noreply.github.com>