Azure App Configuration - Update sample to use configuration provider to consume feature flag and secret reference (#35740)
### Packages impacted by this PR


### Issues associated with this PR


### Describe the problem that is addressed by this PR

The Azure App Configuration team developed [configuration
provider](https://learn.microsoft.com/en-us/azure/azure-app-configuration/configuration-provider-overview)
library. It is built on the azure sdk. The Azure App Configuration team
really wants the customers to use the configuration provider library.

For a long time, we did not provide a configuration provider library for
JavaScript development. As a result, a large number of users have been
relying on the Azure JS SDK, which is a problem for us. Many features,
such as feature flags and secret references, are not suitable for direct
usage through the SDK. The samples for feature flags and secret
references in the SDK repo serve as misleading examples. This has even
polluted the answers generated by current AI models.

### What are the possible designs available to address the problem? If
there are more than one possible design, why was the one in this PR
chosen?


### Are there test cases added in this PR? _(If not, why?)_


### Provide a list of related PRs _(if any)_


### Command used to generate this PR:**_(Applicable only to SDK release
request PRs)_

### Checklists
- [ ] Added impacted package name to the issue description
- [ ] Does this PR needs any fixes in the SDK Generator?** _(If so,
create an Issue in the
[Autorest/typescript](https://github.com/Azure/autorest.typescript)
repository and link it here)_
- [ ] Added a changelog (if necessary)