[App Configuration] - Make list label test deterministic (#35952)
### Packages impacted by this PR
`@azure/app-configuration`

### Issues associated with this PR

#35887

### Describe the problem that is addressed by this PR
parallel CI jobs (or other tests) that introduce an unlabeled setting
can cause an additional { name: null } label to appear between
successive enumerations in `toSortedLabelsArray`, leading to a mismatch.

### What are the possible designs available to address the problem? If
there are more than one possible design, why was the one in this PR
chosen?


### Are there test cases added in this PR? _(If not, why?)_

Yes.

### Provide a list of related PRs _(if any)_


### Command used to generate this PR:**_(Applicable only to SDK release
request PRs)_

### Checklists
- [x] Added impacted package name to the issue description
- [ ] Does this PR needs any fixes in the SDK Generator?** _(If so,
create an Issue in the
[Autorest/typescript](https://github.com/Azure/autorest.typescript)
repository and link it here)_
- [ ] Added a changelog (if necessary)