Azure App Configuration - Update package readme (#35734)
### Packages impacted by this PR

@azure/app-configuration

### Issues associated with this PR


### Describe the problem that is addressed by this PR

Advertise configuration provider in the readme to ensure our customers
use the correct package for different use cases.

The @azure/app-configuration (sdk) package is low level control plane
operation.

The @azure/app-configuration-provider (configuration provider) package
is the client library that most customers should use to consume key
values/feature flags from Azure App Configuration. The configuration
provider package has dependency on the sdk package.

This is a historical issue we have been facing for very long time. Most
of our customers are sticky with the sdk. But the Azure App
Configuration team really want people to use the configuration provider
package. AI's answers usually mention the sdk package instead of the
configuration provider, which makes it even worse. We want to solve this
problem.

### What are the possible designs available to address the problem? If
there are more than one possible design, why was the one in this PR
chosen?


### Are there test cases added in this PR? _(If not, why?)_


### Provide a list of related PRs _(if any)_


### Command used to generate this PR:**_(Applicable only to SDK release
request PRs)_

### Checklists
- [ ] Added impacted package name to the issue description
- [ ] Does this PR needs any fixes in the SDK Generator?** _(If so,
create an Issue in the
[Autorest/typescript](https://github.com/Azure/autorest.typescript)
repository and link it here)_
- [ ] Added a changelog (if necessary)