fix-ppaf-enablement (#36090)
### Packages impacted by this PR
@azure/cosmos

### Issues associated with this PR
https://github.com/Azure/azure-sdk-for-js/issues/35967

### Describe the problem that is addressed by this PR
This PR is follow up of
https://github.com/Azure/azure-sdk-for-js/pull/35954 for fixing the
default value of ppaf flags.
This PR makes the following changes:
1. Always check if PPAF logic is enabled through GlobalEndpointManager
(It contains the most recent value of PPAF settings in account). It is
done this way to account for upcoming dynamic PPAF enablement.
2. Remove PPAF feature usage tracking with useragent for now (need to
come up with a better way).
3. Disable PPAF if enableEndpointDiscovery is false.
4. Code Cleanup

### What are the possible designs available to address the problem? If
there are more than one possible design, why was the one in this PR
chosen?


### Are there test cases added in this PR? _(If not, why?)_
N/A

### Provide a list of related PRs _(if any)_


### Command used to generate this PR:**_(Applicable only to SDK release
request PRs)_

### Checklists
- [ ] Added impacted package name to the issue description
- [ ] Does this PR needs any fixes in the SDK Generator?** _(If so,
create an Issue in the
[Autorest/typescript](https://github.com/Azure/autorest.typescript)
repository and link it here)_
- [ ] Added a changelog (if necessary)