[@azure/cosmos] Add AAD Scope Override and fallback (#36024)
### Packages impacted by this PR
@azure/cosmos

### Issues associated with this PR
https://github.com/Azure/azure-sdk-for-js/issues/36015

### Describe the problem that is addressed by this PR
Added support for overriding AAD authentication scope via the new
`aadScope` option in `CosmosClientOptions`. When no custom scope is
provided, the system uses the account-specific scope for authentication
and implements a fallback mechanism to
`https://cosmos.azure.com/.default` in case of `AADSTS500011` errors.
When a custom scope is explicitly provided via the `aadScope` option, no
fallback occurs.

### What are the possible designs available to address the problem? If
there are more than one possible design, why was the one in this PR
chosen?


### Are there test cases added in this PR? _(If not, why?)_
Yes

### Provide a list of related PRs _(if any)_


### Command used to generate this PR:**_(Applicable only to SDK release
request PRs)_

### Checklists
- [ ] Added impacted package name to the issue description
- [ ] Does this PR needs any fixes in the SDK Generator?** _(If so,
create an Issue in the
[Autorest/typescript](https://github.com/Azure/autorest.typescript)
repository and link it here)_
- [ ] Added a changelog (if necessary)