[attestation] fix build error introduced ty TypeScript 5.9 upgrade (#36180)
After the upgrades, we are getting the following error when building
samples

>@azure/attestation:build:samples:
samples-dev/attestEnclaves.ts(402,18): error TS2352: Conversion of type
'ArrayBuffer' to type 'Buffer<ArrayBufferLike>' may be a mistake because
neither type sufficiently overlaps with the other. If this was
intentional, convert the expression to 'unknown' first.
@azure/attestation:build:samples: Type 'ArrayBuffer' is missing the
following properties from type 'Buffer<ArrayBufferLike>': subarray,
write, toJSON, equals, and 94 more.
@azure/attestation:build:samples: samples-dev/attestEnclaves.ts(410,17):
error TS2345: Argument of type 'Uint8Array<ArrayBufferLike>' is not
assignable to parameter of type 'ArrayBuffer'.
@azure/attestation:build:samples: The types of
'slice(...)[Symbol.toStringTag]' are incompatible between these types.
@azure/attestation:build:samples: Type '"Uint8Array"' is not assignable
to type '"ArrayBuffer"'.

This PR simplifies the sample to just use `TextDecoder`, which is
available in global namespace in our supported NodeJS versions.