[Azure App Configuration] - Normalize the query parameters in request url (#35962)
### Packages impacted by this PR
@azure/app-configuration 

### Issues associated with this PR


### Describe the problem that is addressed by this PR

Adds a new pipeline to update the query params to make sure they are all
lowercase and in alphabetical order. This is required for support of
Azure Front Door as a CDN.

### What are the possible designs available to address the problem? If
there are more than one possible design, why was the one in this PR
chosen?


### Are there test cases added in this PR? _(If not, why?)_

Yes

### Provide a list of related PRs _(if any)_


### Command used to generate this PR:**_(Applicable only to SDK release
request PRs)_

### Checklists
- [x] Added impacted package name to the issue description
- [ ] Does this PR needs any fixes in the SDK Generator?** _(If so,
create an Issue in the
[Autorest/typescript](https://github.com/Azure/autorest.typescript)
repository and link it here)_
- [x] Added a changelog (if necessary)