Add Env Setup to Copilot Instructions (#36533)
https://github.com/Azure/azure-sdk-tools/issues/12592

- Adding instructions on using new Verify Setup MCP tool, which checks
that requirements are installed in the environment
- We want this check to be proactive and happen without explicit
invokation. However, currently, Copilot would only remember if the tool
is run in the current session, not in newer sessions, so the user would
likely run verify setup in each new session due to a lack of a caching.
This will be accepted for Scenario 1 and improved in future iterations

**Example of implicitly running this tool before running a different
requested tool**

<img width="247" height="290" alt="image"
src="https://github.com/user-attachments/assets/86a9da59-6619-4b78-90d8-6c658c29665c"
/>