[test] fix eventHubConsumerClientUnitTests compile error
This error exists and only surfaced recently because of possible improvements in
latest vitest version.

The problem is that `_createEventProcessor` is private in
`EventHubConsumerClient` but is public in the intersection type, leads to a
resolved type of `never` for the first parameter of `spyOn`.  Previous versions
of vitest didn't report on this.

This PR casts `client` to `any` instead, which should be fine in tests.