[App Configuration] - Add audience error handling policy (#36573)
### Packages impacted by this PR
@azure/app-configuration

### Issues associated with this PR


### Describe the problem that is addressed by this PR

Users running in clouds other than the public cloud must correctly
configure ConfigurationClientOptions.Audience when using the
ConfigurationClient from the Azure SDK.

There are two main ways users can misconfigure
They do not specify audience when they are running in non-public cloud
They specify audience, and the audience they specify is the wrong one
for the cloud is using

In both cases we will get an error when trying to get an Entra ID token
that looks like:

"Microsoft.Identity.Client.MsalServiceException: AADSTS500011: The
resource principal named https://appconfig.azure.com/ was not found in
the tenant named msazurecloud. This can happen if the application has
not been installed by the administrator of the tenant or consented to by
any user in the tenant."

We should handle this error and surface up an improved error message

Audience not provided
If we get this error and audience is not provided we should surface an
error message that says audience should be configured and link to our
public doc that documents the appropriate audience for each cloud

Audience provided and incorrect
If we get this error and the audience is provided but is wrong, we
should surface an error message that the configured audience is wrong
and link to our public doc that documents the appropriate audience for
each cloud.




### What are the possible designs available to address the problem? If
there are more than one possible design, why was the one in this PR
chosen?


### Are there test cases added in this PR? _(If not, why?)_


### Provide a list of related PRs _(if any)_


### Command used to generate this PR:**_(Applicable only to SDK release
request PRs)_

### Checklists
- [ ] Added impacted package name to the issue description
- [ ] Does this PR needs any fixes in the SDK Generator?** _(If so,
create an Issue in the
[Autorest/typescript](https://github.com/Azure/autorest.typescript)
repository and link it here)_
- [ ] Added a changelog (if necessary)