[@azure/web-pubsub-client] Use ping/pong as an accurate keepalive (#36751)
### Packages impacted by this PR
@azure/web-pubsub-client

### Issues associated with this PR


### Describe the problem that is addressed by this PR
Web PubSub follows the WebSocket (WS) ping-pong spec and periodically
sends a ping frame to the client, the client should respond with a pong
frame. This mechanism enables Web PubSub to detect client liveness
quickly. But sometimes when there is network interruptions, Web PubSub
detects the client offline quickly, however, client lacks a way to
detect if itself is offline, client might consider itself as still
online for quite some time and fails to reconnect quickly. In some
clients, for example, in the browser, WebSocket ping/pong frames are
quite low-level, and the client does not have access to there ping/pong
frames or events. Here we provide a ping-pong mechanism in the
application layer, providing the clients a way to detect the client
liveness.

### What are the possible designs available to address the problem? If
there are more than one possible design, why was the one in this PR
chosen?


### Are there test cases added in this PR? _(If not, why?)_


### Provide a list of related PRs _(if any)_


### Command used to generate this PR:**_(Applicable only to SDK release
request PRs)_

### Checklists
- [ ] Added impacted package name to the issue description
- [ ] Does this PR needs any fixes in the SDK Generator?** _(If so,
create an Issue in the
[Autorest/typescript](https://github.com/Azure/autorest.typescript)
repository and link it here)_
- [ ] Added a changelog (if necessary)