[Engsys] update analyze deps tool (#36866)
This PR updates the `eng/tool/analyze-deps` tool. The tool is
responsible for generating data of the dependency reports linked at the
top of our repo top level README

-
https://azuresdkartifacts.blob.core.windows.net/azure-sdk-for-js/dependencies/dependencies.html
-
https://azuresdkartifacts.blob.core.windows.net/azure-sdk-for-js/dependencies/dependencyGraph/dagre.html

The first file is still generated properly. The second one hasn't been
generated for quite a while, likely due to pnpm-lock file format changes
and is not working today after we migrate to PNPM. This PR fixes it.