[Identity] Update MI Response Handler in IdentityClient (#36906)
### Packages impacted by this PR
@azure/identity

### Issues associated with this PR

### Describe the problem that is addressed by this PR
- In MSAL implementation of `INetworkModule`, the `HttpClient` will
parse response with no body using `parseBody` and return an object with
an error message. Our `IdentityClient` implementation returns an
undefined body when response body is empty; thus raising `Cannot read
properties of undefined (reading 'expires_on')` when MSAL handles the
response. This PR aligns our behavior according to MSAL `parseBody` so
that MSAL can surface the corresponding error. We also add additional
logging for easier debugging when response body is empty

### What are the possible designs available to address the problem? If
there are more than one possible design, why was the one in this PR
chosen?


### Are there test cases added in this PR? _(If not, why?)_


### Provide a list of related PRs _(if any)_


### Command used to generate this PR:**_(Applicable only to SDK release
request PRs)_

### Checklists
- [ ] Added impacted package name to the issue description
- [ ] Does this PR needs any fixes in the SDK Generator?** _(If so,
create an Issue in the
[Autorest/typescript](https://github.com/Azure/autorest.typescript)
repository and link it here)_
- [ ] Added a changelog (if necessary)