refactor msalClient to export a context and flows directly
Exporting one client with all flows in it as a function makes it
inherently not tree shakable increasing bundle size. This commit
splits all those flows into functions taking the context as a
parameter.

This resulted in a bundle size reduction from 790kb -> 753kb on
an artifical test bundle with only ClientCertificateCredential